requirements.txt
asgiref==3.5.0
Django==3.2.7
mysqlclient==2.0.3
numpy==1.22.0
pandas==1.3.5
pymongo==4.0.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
django-extensions==3.1.3
django-environ==0.7.0
django-filter==2.4.0
django-storages==1.12.2
djangorestframework==3.12.4
google-api-core==2.1.1
google-auth==2.3.0
google-cloud-core==2.1.0
google-cloud-secret-manager==2.7.2
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.0
idna==3.3
libcst==0.3.21
mypy-extensions==0.4.3
mysqlclient==2.0.3
packaging==21.0
Pillow==8.3.2
proto-plus==1.19.5
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
uWSGI==2.0.18